% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coordesc.R
\name{CoorDesc}
\alias{CoorDesc}
\title{This function implements the coordinate-wise gradient boosting algorithm}
\usage{
CoorDesc(delta, z, time, p, N, knot, maxit = 300, rate = 0.05,
  tol = 0.001)
}
\arguments{
\item{delta}{event indicator}

\item{z}{Covariate matrix}

\item{time}{the death time}

\item{p}{number of variables}

\item{N}{the sample size}

\item{knot}{the number of basis functions for non-linear predictors}

\item{maxit}{the number of maximum iterations}

\item{rate}{the learning rate}

\item{tol}{the convergence threshold}
}
\value{
return a list of the following values
beta :  1 denotes selected variables and 0 denotes unselected variables.
time :  the used time
convergence : whether it's converge within maximum iteration steps
key  :  the number of iterations
}
\description{
This function implements the coordinate-wise gradient boosting algorithm
}
