\name{CoxAdditive-package}
\alias{CoxAdditive-package}
\alias{CoxAdditive}
\docType{package}
\title{\packageTitle{CoxAdditive}}
\description{\packageDescription{CoxAdditive}}
\details{
  \packageIndices{CoxAdditive}

  This package provides a simulating function, and two main functions implementing coordinate-wise gradient ascent and mmboosting algorithm.
  First use simulation function to get simulated data. Then put the data to main functions.
}
\author{
  \packageAuthor{CoxAdditive}
  
  Maintainer: \packageMaintainer{CoxAdditive}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{

}
