% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ddloglik}
\alias{ddloglik}
\title{An Rcpp function that calculates the first and second order derivate and partial likelihood function.}
\usage{
ddloglik(n, delta, z, beta, offset)
}
\arguments{
\item{delta}{Event indicator}

\item{z}{Covariate matrix}

\item{beta}{The coefficient}

\item{offset}{The whole spanned matrix}

\item{N}{The sample size}
}
\value{
the first and second derivate and partial likelihood
}
\description{
An Rcpp function that calculates the first and second order derivate and partial likelihood function.
}
