% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coxsrc.R
\name{simul}
\alias{simul}
\title{Simulation function
the function for simulating data}
\usage{
simul(N = 1000, p = 500, p_true = 5, m1 = 100, knot = 10)
}
\arguments{
\item{N}{sample size}

\item{p}{number of parameters}

\item{p_true}{number of true parameters}
}
\value{
a list of the following values

delta: event indicator
z: Covariate matrix
time: the death time
p: number of variables
N: the sample size
knot: the number of basis functions for non-linear predictors
}
\description{
Simulation function
the function for simulating data
}
