% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coxsrc.R
\name{FPFNSeSpLik}
\alias{FPFNSeSpLik}
\title{The evaluation function}
\usage{
FPFNSeSpLik(TrueBeta = TrueBeta, beta = beta)
}
\value{
a list of the following values
FP:  false positive 
FN:  false negative
Se:  Senstitivity
Sp:  Specificity
FDP: False positive portion
}
\description{
The evaluation function
}
